Make uniform-varint-length random number generator available to benches.
I want to use that in interval log benchmarks, hence tackling that TODO.

For serialization benchmarks, this is just right: we want to exercise
all different varint lengths (modulo bucketing, which throws it off a
little bit). For non-serialization benches (e.g. add and subtract) this
is still a lot better than just regular random numbers, which will be
heavily weighted towards the last few buckets.

This drops measurements quite a bit, but that's expected because now
histograms aren't mostly using values way in the high end that all get
bucketed together, so now there's a lot more work to do when traversing
the histogram.
